refactor(gax): move `ExponentialBackoff`
I am planning to use the same implementation (with a different trait) in
the policies controlling the LRO polling loop.
