+++ src/gax/src/lib.rs
@@ -82,6 +82,7 @@ pub mod paginator;
 pub mod http_client;
 
 pub mod backoff_policy;
+pub mod exponential_backoff;
 pub mod options;
 pub mod retry_policy;
 pub mod retry_throttler;
